Update MD_MIDITrack.cpp
Just proposing this change, i'm not sure that it won't negatively impact other library users.....

This fix is for the ESP32 to work with the library. The ESP32 developers thought it would be a good idea to eschew the use of the min() macro provided by the standardized Arduino API in favor of the template based version used in the STL. Unlike Arduino's min(), this version of min() does not accept comparisons between different types.